tests/krate/yanking: Move route tests into `routes` submodules
