+++ src/tests/routes/crates/versions/mod.rs
@@ -2,3 +2,4 @@ mod authors;
 pub mod dependencies;
 pub mod download;
 mod read;
+pub mod yank_unyank;
